+++ lib/simple_flow.rb
@@ -19,6 +19,7 @@ require_relative 'simple_flow/result'
 require_relative 'simple_flow/pipeline'
 require_relative 'simple_flow/middleware'
 require_relative 'simple_flow/step_tracker'
+require_relative 'simple_flow/parallel_step'
 
 module SimpleFlow
 end